Improve server mesh performance

+++ src/main/java/com/github/argon4w/acceleratedrendering/core/meshes/ClientMesh.java
@@ -84,7 +84,7 @@ public class ClientMesh implements IMesh {
 
 			builders.add(builder);
 
-			mesh = new ClientMesh(vertexCount, result.byteBuffer());
+			mesh = new ClientMesh((int) vertexCount, result.byteBuffer());
 
 			MeshDataCaches.CLIENT.set(
 					layout,
