+++ src/main/java/com/github/argon4w/acceleratedrendering/core/meshes/collectors/CulledMeshCollector.java
@@ -154,7 +154,7 @@ public class CulledMeshCollector implements IMeshCollector {
 	}
 
 	@Override
-	public int getVertexCount() {
+	public long getVertexCount() {
 		return meshCollector.getVertexCount();
 	}
 }
